feat : HeaderService 컴포넌트 추가 및 RollingPaper폴더 추가
