+++ .env.development
@@ -1,2 +1,3 @@
 # 개발 환경용 환경 변수
-VITE_KAKAO_APP_KEY=your_kakao_app_key_here
+VITE_KAKAO_APP_KEY=5c30f812d043089f85e15542f84eb0f7
+
