+++ src/pages/RollingPaper/components/MessageItem.jsx
@@ -1,8 +1,8 @@
-import Avatar from "./Avatar";
-import Badge from "./Badge";
-import { formatDate } from "../../utils/utils";
+import Avatar from "@/components/ui/Avatar";
+import Badge from "@/components/ui/Badge";
+import { formatDate } from "@/utils/utils";
 import "./MessageItem.scss";
-import AddIcon from "../../assets/images/icons/add.svg";
+import AddIcon from "@/assets/images/icons/add.svg";
 
 /**
  * 메시지 카드 컴포넌트 - 일반 메시지 표시 또는 새 메시지 추가 카드
