+++ src/pages/Test/TestColor/style.scss
@@ -21,8 +21,7 @@ body {
   font-weight: 600;
   color: #333;
 }
-.group {
-}
+
 h3 {
   margin: 0 0 10px;
   font-size: 14px;
