+++ src/pages/Test/TestComponents2/style.scss
@@ -1,23 +1,5 @@
-/* 스피너 애니메이션 */
-@keyframes spin {
-  0% {
-    transform: rotate(0deg);
-  }
-  100% {
-    transform: rotate(360deg);
-  }
-}
-
 /* Page Wrapper - 다른 페이지에 영향 없이 이 페이지만 배경 적용 */
 .test2-page-wrapper {
-  /* CSS 변수를 사용한 조건부 배경 설정 */
-  background-color: var(--bg-color, var(--c-beige200));
-  background-image: var(--bg-image, none);
-  background-size: cover;
-  background-position: center center;
-  background-repeat: no-repeat;
-  background-attachment: scroll;
-
   position: relative;
   width: 100vw;
   min-height: 100vh;
@@ -26,225 +8,26 @@
   overflow-x: hidden; /* 가로 스크롤만 제거 */
   padding: 0;
   box-sizing: border-box;
-  padding-top: 68px; /* 고정 헤더 높이만큼 여백 */
+  padding-top: 68px; /* 메인 헤더(64px) + 서비스 헤더(68px) */
 
   /* 추가 가로 스크롤 방지 */
   max-width: 100vw;
 }
 
-/* Header Service */
-.header-service {
-  position: fixed; /* 완전 고정 */
-  top: 64px; /* 상단에 고정 */
-  left: 0;
-  right: 0;
-  background: white;
-  border-bottom: 1px solid #eee;
-  height: 68px; /* 고정 높이 */
-  display: flex;
-  align-items: center;
-  z-index: 1000; /* 다른 요소들 위에 표시 */
-
-  .header-content {
-    max-width: 1200px;
-    margin: 0 auto;
-    padding: 0 24px;
-    display: flex;
-    align-items: center;
-    gap: 16px;
-    width: 100%;
-
-    h1 {
-      font-size: 28px;
-      font-weight: 700;
-      margin: 0;
-      text-align: left;
-    }
-
-    p {
-      margin-left: 10px;
-      font-size: 18px;
-      color: #333;
-      font-weight: 400;
-    }
-
-    .divider {
-      background-color: var(--c-gray200, #ddd);
-      width: 1px;
-      height: 28px;
-    }
-
-    .emoji-group {
-      position: relative;
-      display: flex;
-      align-items: center;
-      gap: 8px;
-
-      .reactions-container {
-        display: flex;
-        align-items: center;
-        gap: 8px;
-      }
-
-      .emoji-dropdown-toggle {
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        padding: 8px 12px;
-        color: var(--c-gray400);
-        cursor: pointer;
-        transition: all 0.2s ease;
-        font-size: 16px;
-      }
-    }
-
-    .emoji-add-button,
-    .share-button {
-      padding: 12px 16px;
-      border: 1px solid #ddd;
-      border-radius: 6px;
-      background: white;
-      font-size: 16px;
-      cursor: pointer;
-      transition: all 0.2s ease;
-
-      &:hover {
-        background: #f5f5f5;
-        border-color: #999;
-      }
-    }
-
-    .share-button {
-      background: #007bff;
-      color: white;
-      border-color: #007bff;
-
-      &:hover {
-        background: #0056b3;
-        border-color: #0056b3;
-      }
-    }
-  }
-}
-
 /* Message List Wrapper - 배경 이미지와 스크롤 처리 */
 .message-list-wrapper {
   /* CSS 변수를 사용한 조건부 배경 설정 */
-  background-color: var(--bg-color, #ffe2ad);
+  background-color: var(--bg-color, var(--c-beige200));
   background-image: var(--bg-image, none);
   background-size: cover;
   background-position: center center;
   background-repeat: no-repeat;
-  background-attachment: fixed;
-
+  //background-attachment: fixed;
+  padding-top: 113px;
   overflow: visible; /* 스크롤 제거 */
-  padding: 40px 0; /* 세로 패딩만 유지 */
 
-  // Tablet 스타일 (769px ~ 1247px)
-  @media (max-width: 1247px) and (min-width: 769px) {
-    padding: 40px 0; /* 세로 패딩 유지 */
-  }
-
-  // Mobile 스타일 (768px 이하)
-  @media (max-width: 768px) {
-    padding: 40px 0; /* 세로 패딩 유지 */
-  }
-}
-
-/* 이모지 드롭다운 스타일 */
-.emoji-dropdown-container {
-  position: absolute;
-  top: calc(100% + 10px);
-  left: 0;
-  background-color: white;
-  border: 1px solid #ddd;
-  border-radius: 12px;
-  padding: 12px;
-  box-shadow: 0 4px 12px rgba(0, 0, 0, 0.1);
-  z-index: 10;
-  width: 280px;
-  box-sizing: border-box;
-
-  .dropdown-reactions {
-    .reactions--badges {
-      display: grid;
-      grid-template-columns: repeat(4, 1fr);
-      gap: 6px;
-    }
-  }
-}
-
-/* 이모지 추가 섹션 스타일 */
-.emoji-add-section {
-  position: relative;
-  display: inline-block;
-}
-
-/* 이모지 피커 드롭다운 스타일 */
-.emoji-picker-dropdown {
-  position: absolute;
-  top: 100%;
-  right: 0; /* 버튼의 오른쪽 끝에 정렬 */
-  z-index: 1000;
-  background: white;
-  border-radius: 12px;
-  box-shadow: 0 8px 32px rgba(0, 0, 0, 0.15);
-  border: 1px solid #e9ecef;
-  overflow: hidden;
-  margin-top: 8px;
-}
-
-/* 공유 섹션 스타일 */
-.share-section {
-  position: relative;
-  display: inline-block;
-}
-
-/* 공유 드롭다운 스타일 */
-.share-dropdown {
-  position: absolute;
-  top: 100%;
-  right: 0;
-  z-index: 1000;
-  background: white;
-  border-radius: 8px;
-  box-shadow: 0 4px 20px rgba(0, 0, 0, 0.15);
-  border: 1px solid #e9ecef;
-  overflow: hidden;
-  margin-top: 8px;
-  min-width: 140px;
-
-  .share-option {
-    display: block;
-    width: 100%;
-    padding: 12px 16px;
-    background: white;
-    border: none;
-    text-align: left;
-    font-size: 14px;
-    font-family: "Pretendard", sans-serif;
-    color: #333;
-    cursor: pointer;
-    transition: background-color 0.2s ease;
-
-    &:hover {
-      background-color: #f8f9fa;
-    }
-
-    &:not(:last-child) {
-      border-bottom: 1px solid #f0f0f0;
-    }
-
-    &:first-child {
-      border-radius: 8px 8px 0 0;
-    }
-
-    &:last-child {
-      border-radius: 0 0 8px 8px;
-    }
-
-    &:only-child {
-      border-radius: 8px;
-    }
-  }
+  // // Mobile 스타일 (768px 이하)
+  // @media (max-width: 768px) {
+  //   padding: 40px 0; /* 세로 패딩 유지 */
+  // }
 }
